md-select-menu should always appear on top of its owner md-select
md-select-menu was ending up with a z-index that is not related to the effective
z-index of the md-select that owns it, so it may end up below the container of
said md-select when it is used inside a $mdPanel or a $mdDialog.
This fix will compute the actual z-index of md-select and it will apply it
to both md-select-menu and md-backdrop.